requirements.txt
flask==3.0.0
streamlit==1.29.0
langchain==0.0.350
openai==1.3.0
pytesseract==0.3.10
easyocr==1.7.1
spacy==3.7.2
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
fpdf==1.7.2
Pillow==10.1.0
python-dotenv==1.0.0
tiktoken==0.5.1
pdf2image==1.16.3
opencv-python-headless==4.8.1.78
